Moved one method from external and fixed some minor things
